Extract tests for isDate to own file
